Inject devices and dictionary data from the top level
Importing JSON files requires a custom import declaration specific of NakedJSX (or probably Rollup). This makes the codebase less reusable and portable across technologies. Furthermore, pure data doesn't belong in the `lib` folder anymore.

The solution is limiting the use of those imports to contexts specific of NakedJSX; that is, pages. After that, JSON files are better placed at a root `data` folder.